Enhancements to the Eiffel generator #6057 Added shell scripts, batch files to generate Eiffel Petstore samples. Added Setup CI (travis, shippable, etc) to cover the Eiffel Petstore samples. Refactored part of the code into an Abstract Eiffel base class, which can be later used in Eiffel server stub generators. Added Eiffel Style guide to the contribution guidelines so that Eiffel contributors know what to follow in terms of code style Fixed code generator issues. Updated mustaches templates.
