+++ modules/swagger-codegen/src/main/resources/Eiffel/api_client.mustache
@@ -1,4 +1,4 @@
-{{>noteInfo}}  
+{{>noteinfo}}  
 
 class
 	API_CLIENT
