+++ modules/swagger-codegen/src/main/resources/Eiffel/framework/auth/authentication.mustache
@@ -1,4 +1,4 @@
-{{>noteInfo}}  
+{{>noteinfo}}  
 
 deferred class
 	AUTHENTICATION
