+++ modules/swagger-codegen/src/main/resources/Eiffel/travis.mustache
@@ -1,0 +1,20 @@
+language: eiffel
+before_script:
+  - export current_dir=$(pwd)
+  - echo current_dir
+  - cd ..
+  - wget https://ftp.eiffel.com/pub/beta/nightly/Eiffel_17.11_gpl_100608-linux-x86-64.tar.bz2  
+  - tar -xvf Eiffel_17.11_gpl_100608-linux-x86-64.tar.bz2
+  - export ISE_EIFFEL=$PWD/Eiffel_17.11
+  - export ISE_PLATFORM=linux-x86-64
+  - export PATH=$PATH:$ISE_EIFFEL/studio/spec/$ISE_PLATFORM/bin
+  - export PATH=$PATH:$ISE_EIFFEL/tools/spec/$ISE_PLATFORM/bin
+  - cd $current_dir
+
+# safelist
+branches:
+  only:
+  - master
+
+script: compile_all -ecb -melt -list_failures -clean -options dotnet=false
+
